fix the pro-rataed leave with inactive and with new resource joined

+++ src/main/java/com/example/leavemanagement/dto/EmployeeLeaveDetail.java
@@ -33,7 +33,7 @@ public record EmployeeLeaveDetail(
         int quarter,
         LocalDate quarterStart,
         LocalDate quarterEnd,
-        int permissibleLeave,
+        double permissibleLeave,
         int carriedForwardLeave,
         double leaveTaken,
         double paidLeave,
